requirements.txt
streamlit==1.31.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.18.0
Pillow==10.2.0